modif route DELETE

+++ back/controllers/sauce.js
@@ -10,7 +10,7 @@ exports.createSauce = (req, res, next) => {
         userId: req.auth.userId,
         imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
     });
-    thing.save()
+    sauce.save()
     .then(() => { res.status(201).json({message: 'Sauce enregistrée !'})})
     .catch(error => { res.status(400).json( { error })})
     }
@@ -23,11 +23,11 @@ exports.modifySauce = (req, res, next) => {
 
     delete sauceObject._userId;
     Sauce.findOne({_id: req.params.id})
-    .then((thing) => {
-        if (thing.userId != req.auth.userId) {
+    .then((sauce) => {
+        if (sauce.userId != req.auth.userId) {
             res.status(401).json({ message: 'non autorisé'});
         } else {
-            Sauce.updateOne({ _id: req.params.id}, { ...thingObject, _id: req.params.id})
+            Sauce.updateOne({ _id: req.params.id}, { ...sauceObject, _id: req.params.id})
             .then(() => res.status(200).json({message: 'Sauce modifiée'}))
             .catch(error => res.status(401).json({ error }))
         }
@@ -38,10 +38,23 @@ exports.modifySauce = (req, res, next) => {
     }
 
 exports.deleteSauce = (req, res, next) => {
-    Sauce.deleteOne({ _id: req.params.id })
-        .then(() => res.status(200).json({ message: 'Sauce supprimée !'}))
-        .catch(error => res.status(400).json({ error }))
-    }
+    Sauce.findOne({ _id: req.params.id})
+        .then(sauce => {
+            if (sauce.userId != req.auth.userId) {
+                res.status(401).json({message: 'Not authorized'});
+            } else {
+                const filename = sauce.imageUrl.split('/images/')[1];
+                fs.unlink(`images/${filename}`, () => {
+                    Sauce.deleteOne({_id: req.params.id})
+                        .then(() => { res.status(200).json({message: 'Objet supprimé !'})})
+                        .catch(error => res.status(401).json({ error }));
+               });
+           }
+        })
+        .catch( error => {
+            res.status(500).json({ error });
+        });
+};
 
 exports.getOneSauce = (req, res, next) => {
     Sauce.findOne({ _id: req.params.id })
